requirements.txt
langchain 
wikipedia 
langchain-google-genai
langchain-community 
python-dotenv
pydantic
duckduckgo-search

